Added buy and sell logic

+++ apps/register/models.py
@@ -22,12 +22,12 @@ class UserManager(models.Manager):
 
 class User(models.Model):
     name = models.CharField(max_length=255)
-    deposit_amount = models.IntegerField()
+    deposit_amount = models.DecimalField(max_digits=15, decimal_places=2)
     created_at = models.DateTimeField(auto_now_add = True)
     updated_at = models.DateTimeField(auto_now = True)
 
 class VGC(models.Model):
-    amount = models.IntegerField()
+    amount = models.DecimalField(max_digits=15, decimal_places=2)
     price = models.DecimalField(max_digits=10, decimal_places=2)
     created_at = models.DateTimeField(auto_now_add = True)
     updated_at = models.DateTimeField(auto_now = True)
